kept task 3 only from shuvo's branch into this file. Added file named pysh_ens_variance_virag for the function
